refactor: components refactor and added notfound page + footer
