--- src/components/Footer.tsx
@@ -1,12 +1,0 @@
-export default function Footer() {
-  return (
-    <footer className="flex justify-between items-center bg-gray-800 p-4">
-      <ul>
-        <li>About</li>
-        <li>Contact</li>
-        <li>Privacy Policy</li>
-        <li>Terms of Service</li>
-      </ul>
-    </footer>
-  );
-}
