+++ src/components/explore/ExplorePosts.tsx
@@ -2,12 +2,12 @@ import PostCard from "../posts/PostCard";
 
 export default function ExplorePosts() {
   return (
-    <>
+    <section className="relative border-x-[1px] border-tertiary-500">
       <PostCard />
       <PostCard image={true} />
       <PostCard image={true} />
       <PostCard />
       <PostCard />
-    </>
+    </section>
   );
 }
