+++ src/components/profile/ProfileBanner.tsx
@@ -1,0 +1,10 @@
+export default function ProfileBanner() {
+  return (
+    <div>
+      <img
+        src="/placeholder_horizontal.jpg"
+        className="aspect-[3/1] w-full object-cover object-center"
+      ></img>
+    </div>
+  );
+}
