+++ src/components/search/SearchAll.tsx
@@ -1,59 +1,23 @@
-import Avatar from "../ui/Avatar";
-import Divider from "../ui/Divider";
-import FollowBtn from "../ui/FollowBtn";
-import NickName from "../userPanel/NickName";
+import UserProfile from "../userPanel/UserProfile";
 import Hashtag from "../ui/Hashtag";
 import PostCard from "../posts/PostCard";
 
 export default function SearchAll() {
   return (
-    <>
+    <section className="relative border-x-[1px] border-tertiary-500">
       <div>
         <p className="border-b border-tertiary-600 bg-tertiary-600/25 p-5 text-xs font-bold uppercase tracking-wider text-tertiary-200">
           Profiles
         </p>
-        <div className="flex items-center justify-between p-5">
-          <div className="flex cursor-pointer items-center gap-2.5">
-            <div className="max-w-12">
-              <Avatar
-                src={"/avatar_placeholder.jpg"}
-                alt="Avatar placeholder"
-                indicator={false}
-              />
-            </div>
-            <NickName nickname="nickname" />
-          </div>
-          <FollowBtn isFollowing={false} />
-        </div>
-        <Divider />
-        <div className="flex items-center justify-between p-5">
-          <div className="flex cursor-pointer items-center gap-2.5">
-            <div className="max-w-12">
-              <Avatar
-                src={"/avatar_placeholder.jpg"}
-                alt="Avatar placeholder"
-                indicator={false}
-              />
-            </div>
-            <NickName nickname="nickname" />
-          </div>
-          <FollowBtn isFollowing={true} />
-        </div>
-        <Divider />
-        <div className="flex items-center justify-between p-5">
-          <div className="flex cursor-pointer items-center gap-2.5">
-            <div className="max-w-12">
-              <Avatar
-                src={"/avatar_placeholder.jpg"}
-                alt="Avatar placeholder"
-                indicator={false}
-              />
-            </div>
-            <NickName nickname="nickname" />
-          </div>
-          <FollowBtn isFollowing={true} />
-        </div>
-        <Divider />
+        <article className="border-y border-tertiary-500 p-5">
+          <UserProfile isUserPanel={false} isMainUser={false} />
+        </article>
+        <article className="border-y border-tertiary-500 p-5">
+          <UserProfile isUserPanel={false} isMainUser={false} />
+        </article>
+        <article className="border-y border-tertiary-500 p-5">
+          <UserProfile isUserPanel={false} isMainUser={false} />
+        </article>
       </div>
       <div>
         <p className="border-b border-tertiary-600 bg-tertiary-600/25 p-5 text-xs font-bold uppercase tracking-wider text-tertiary-200">
@@ -71,6 +35,6 @@ export default function SearchAll() {
         <PostCard image={true} />
         <PostCard image={true} />
       </div>
-    </>
+    </section>
   );
 }
