+++ src/components/ui/FollowBtn.tsx
@@ -1,6 +1,8 @@
 export default function FollowBtn({ isFollowing }: { isFollowing: boolean }) {
   return (
-    <button className={`${isFollowing ? "btn-secondary" : "btn-primary"}`}>
+    <button
+      className={`${isFollowing ? "btn-secondary" : "btn-primary"} text-sm font-semibold`}
+    >
       <span>{isFollowing ? "Unfollow" : "Follow"}</span>
     </button>
   );
