+++ src/components/userPanel/UserProfile.tsx
@@ -2,16 +2,21 @@ import { Link } from "react-router-dom";
 import Avatar from "../ui/Avatar";
 import Dropdown from "../ui/Dropdown";
 import NickName from "./NickName";
+import FollowBtn from "../ui/FollowBtn";
 
 interface UserProfileProps {
-  isUserPanel?: boolean;
+  isUserPanel: boolean;
+  isMainUser: boolean;
 }
 
-export default function UserProfile({ isUserPanel = false }: UserProfileProps) {
+export default function UserProfile({
+  isUserPanel = false,
+  isMainUser = false,
+}: UserProfileProps) {
   const nickname = "nicknameee";
   return (
     <div
-      className={`flex cursor-pointer items-center justify-between gap-2.5 rounded-lg ${isUserPanel ? "border-2 border-tertiary-400 bg-tertiary-500 p-2.5" : "bg-transparent p-0"} `}
+      className={`flex items-center justify-between gap-2.5 rounded-lg ${isUserPanel ? "border-2 border-tertiary-400 bg-tertiary-500 p-2.5" : "bg-transparent p-0"} `}
     >
       <Link to={`/profile/${nickname}`}>
         <div className="group flex items-center gap-2.5">
@@ -19,13 +24,13 @@ export default function UserProfile({ isUserPanel = false }: UserProfileProps) {
             <Avatar
               src={"/avatar_placeholder.jpg"}
               alt="Avatar placeholder"
-              indicator={true}
+              indicator={isUserPanel}
             />
           </div>
           <NickName nickname="nickname" />
         </div>
       </Link>
-      <Dropdown />
+      {isMainUser ? <Dropdown /> : <FollowBtn isFollowing={false} />}
     </div>
   );
 }
